pyproject.toml
[tool.poetry]
name = "ovpn-portal"
version = "0.1.0"
description = ""
authors = ["Giovanni Barillari <gi0baro@d4net.org>"]
license = "BSD-3-Clause"

[tool.poetry.dependencies]
python = "^3.9"
boto3 = "^1.21.44"
emmett = {extras = ["crypto"], version = "~2.4.9"}
gunicorn = "^20.1.0"
httpx = "^0.22.0"
python-jose = "^3.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
